PR-987140: fixing npx.

+++ bin/cli.js
@@ -15,8 +15,13 @@ const runCommand = command => {
 
 const repoName = process.argv[2];
 const gitCheckoutCommand = `git clone --depth 1 https://github.com/aleleba/create-react-ssr ${repoName}`;
+console.log(`Cloning the repository with name ${repoName}`);
+const checkedOut = runCommand(gitCheckoutCommand);
+if(!checkedOut) process.exit(-1);
+
 const actualVersion = runCommand(`cd ${repoName} && node -p "require('./package.json').version"`)
 if(!actualVersion) process.exit(-1);
+
 const installDepsCommand = `cd ${repoName} && npm install`;
 const cleanGitHistoryCommand = `cd ${repoName} && rm -rf .git && git init && git add --all -- ":!.github" ":!bin" && git commit -m "Initial commit"`
 const cleanGitHistoryCommandWindows = `cd ${repoName} && rmdir .git /s /q && git init && git add --all -- ":!.github" ":!bin" && git commit -m "Initial commit"`
@@ -28,10 +33,6 @@ const deleteBinCommandApple = `cd ${repoName} && sed -i .copy 's+"bin": "./bin/c
 rm -rf package.json.copy`
 const replaceNewVersionCommand = `cd ${repoName} && sed -i 's+"version": "${actualVersion}",+"version": "0.0.1",+g' package.json`
 const replaceNameAppCommand = `cd ${repoName} && sed -i 's+"name": "@aleleba/create-react-ssr",+"name": "${repoName}",+g' package.json`
-
-console.log(`Cloning the repository with name ${repoName}`);
-const checkedOut = runCommand(gitCheckoutCommand);
-if(!checkedOut) process.exit(-1);
 
 console.log(`Installing dependencies for ${repoName}`);
 const installedDeps = runCommand(installDepsCommand);
